refactor jobs
