--- teamster/local/config/datagun/template-query-gsheets.yaml
@@ -1,20 +1,0 @@
-execution:
-  config:
-    multiprocess:
-      max_concurrent: 1
-resources:
-  destination:
-    config:
-      folder_id: ""
-ops:
-  compose_queries:
-    config:
-      destination:
-        name: ""
-        type: ""
-      queries:
-        - file:
-            stem: ""
-            suffix: ""
-          sql:
-            text: ""
